requirements.txt
Flask
pytesseract
pdf2image
Pillow